delta of call options
Delta is the option's sensitivity to small changes in the underlying asset price. 
As a call option gets deep-in-the-money, N(d1) approaches 1, but it never exceeds 1 (since it's a cumulative distribution function). For a European call option on a nondividend-paying stock, N(d1) is moreover equal to the option's delta. Delta can therefore never exceed 1 for this option. F